Remove SecurePushSubscription
