+++ src/NuGetGallery/App_Start/DefaultDependenciesModule.cs
@@ -247,12 +247,6 @@ namespace NuGetGallery
                 .As<IApiScopeEvaluator>()
                 .InstancePerLifetimeScope();
 
-            builder.RegisterType<SecurePushSubscription>()
-                .SingleInstance();
-
-            builder.RegisterType<RequireSecurePushForCoOwnersPolicy>()
-                .SingleInstance();
-
             var mailSenderThunk = new Lazy<IMailSender>(
                 () =>
                 {
