--- src/NuGetGallery/Security/SecurePushSubscription.cs
@@ -1,100 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using NuGet.Versioning;
-using NuGetGallery.Authentication;
-using NuGetGallery.Auditing;
-using NuGetGallery.Diagnostics;
-
-namespace NuGetGallery.Security
-{
-    /// <summary>
-    /// User security policies for the secure push subscription.
-    /// </summary>
-    public class SecurePushSubscription : IUserSecurityPolicySubscription
-    {
-        public const string Name = "SecurePush";
-        internal const string MinProtocolVersion = "4.1.0";
-        internal const int PushKeysExpirationInDays = 30;
-
-        private IAuditingService _auditing;
-        private IDiagnosticsSource _diagnostics;
-
-        /// <summary>
-        /// Subscription name.
-        /// </summary>
-        public string SubscriptionName
-        {
-            get
-            {
-                return Name;
-            }
-        }
-
-        /// <summary>
-        /// Required policies for this subscription.
-        /// </summary>
-        public IEnumerable<UserSecurityPolicy> Policies
-        {
-            get
-            {
-                yield return new UserSecurityPolicy(RequirePackageVerifyScopePolicy.PolicyName, SubscriptionName);
-                yield return RequireMinProtocolVersionForPushPolicy.CreatePolicy(SubscriptionName, new NuGetVersion(MinProtocolVersion));
-            }
-        }
-
-        public SecurePushSubscription(IAuditingService auditing, IDiagnosticsService diagnostics)
-        {
-            _auditing = auditing ?? throw new ArgumentNullException(nameof(auditing));
-
-            if (diagnostics == null)
-            {
-                throw new ArgumentNullException(nameof(diagnostics));
-            }
-
-            _diagnostics = diagnostics.SafeGetSource(nameof(SecurePushSubscription));
-        }
-
-        /// <summary>
-        /// On subscribe, set API keys with push capability to expire in <see cref="PushKeysExpirationInDays" /> days.
-        /// </summary>
-        /// <param name="context"></param>
-        public async Task OnSubscribeAsync(UserSecurityPolicySubscriptionContext context)
-        {
-            var pushKeys = context.User.Credentials.Where(c =>
-                c.IsApiKey() &&
-                (
-                    c.Scopes.Count == 0 ||
-                    c.Scopes.Any(s =>
-                        s.AllowedAction.Equals(NuGetScopes.PackagePush, StringComparison.OrdinalIgnoreCase) ||
-                        s.AllowedAction.Equals(NuGetScopes.PackagePushVersion, StringComparison.OrdinalIgnoreCase)
-                        ))
-                );
-
-            var expires = DateTime.UtcNow.AddDays(PushKeysExpirationInDays);
-            var expireTasks = new List<Task>();
-            foreach (var key in pushKeys)
-            {
-                if (!key.Expires.HasValue || key.Expires > expires)
-                {
-                    expireTasks.Add(_auditing.SaveAuditRecordAsync(
-                        new UserAuditRecord(context.User, AuditedUserAction.ExpireCredential, key)));
-
-                    key.Expires = expires;
-                }
-            }
-            await Task.WhenAll(expireTasks);
-            
-            _diagnostics.Information($"Expiring {pushKeys.Count()} keys with push capability.");
-        }
-        
-        public Task OnUnsubscribeAsync(UserSecurityPolicySubscriptionContext context)
-        {
-            return Task.CompletedTask;
-        }
-    }
-}
