+++ tests/NuGetGallery.Facts/Controllers/PackagesControllerFacts.cs
@@ -1136,118 +1136,6 @@ namespace NuGetGallery
                     messageService.Verify(m => m.SendPackageOwnerRequestCancellationNotice(userA, userB, package));
                 }
             }
-
-            public class TheConfirmOwnerMethod_SecurePushPropagation : TestContainer
-            {
-                [Fact]
-                public async Task SubscribesOwnersToSecurePushAndSendsEmailIfNewOwnerRequires()
-                {
-                    // Arrange
-                    var fakes = Get<Fakes>();
-                    fakes.Package.Owners.Add(fakes.ShaUser);
-                    fakes.User.SecurityPolicies = new RequireSecurePushForCoOwnersPolicy().Policies.ToList();
-
-                    Assert.Equal(0, fakes.Owner.SecurityPolicies.Count);
-
-                    // Act & Assert
-                    var policyMessages = await AssertConfirmOwnerSubscribesUser(fakes, fakes.Owner, fakes.ShaUser, fakes.OrganizationOwner);
-                    Assert.Equal(4, policyMessages.Count);
-
-                    // subscribed notification
-                    Assert.StartsWith("Owner(s) 'testUser' has (have) the following requirements that are now enforced for your account:",
-                        policyMessages[fakes.Owner.Username]);
-                    Assert.StartsWith("Owner(s) 'testUser' has (have) the following requirements that are now enforced for your account:",
-                        policyMessages[fakes.ShaUser.Username]);
-                    Assert.StartsWith("Owner(s) 'testUser' has (have) the following requirements that are now enforced for your account:",
-                        policyMessages[fakes.OrganizationOwner.Username]);
-
-                    // propagator notification
-                    Assert.StartsWith("Owner(s) 'testUser' has (have) the following requirements that are now enforced for co-owner(s) '",
-                        policyMessages[fakes.User.Username]);
-                }
-
-                [Fact]
-                public async Task SubscribesNewOwnerToSecurePushAndSendsEmailIfOwnerRequires()
-                {
-                    // Arrange
-                    var fakes = Get<Fakes>();
-                    fakes.Package.Owners.Add(fakes.ShaUser);
-                    fakes.Owner.SecurityPolicies = new RequireSecurePushForCoOwnersPolicy().Policies.ToList();
-
-                    Assert.Equal(0, fakes.User.SecurityPolicies.Count);
-
-                    // Act & Assert
-                    var policyMessages = await AssertConfirmOwnerSubscribesUser(fakes, fakes.User);
-
-                    Assert.False(policyMessages.ContainsKey(fakes.User.Username));
-                    Assert.Equal(3, policyMessages.Count);
-                    Assert.StartsWith("Owner(s) 'testPackageOwner' has (have) the following requirements that are now enforced for co-owner(s) 'testUser':",
-                        policyMessages[fakes.Owner.Username]);
-                    Assert.Equal("", policyMessages[fakes.ShaUser.Username]);
-                    Assert.Equal("", policyMessages[fakes.OrganizationOwner.Username]);
-                }
-
-                private async Task<IDictionary<string, string>> AssertConfirmOwnerSubscribesUser(Fakes fakes, params User[] usersSubscribed)
-                {
-                    // Arrange
-                    var mockHttpContext = new Mock<HttpContextBase>();
-
-                    var packageService = new Mock<IPackageService>();
-                    packageService.Setup(p => p.FindPackageRegistrationById(It.IsAny<string>())).Returns(fakes.Package);
-
-                    var packageOwnershipManagementService = new Mock<IPackageOwnershipManagementService>();
-                    packageOwnershipManagementService.Setup(p => p.GetPackageOwnershipRequest(fakes.Package, fakes.User, "token")).Returns(
-                        new PackageOwnerRequest
-                        {
-                            PackageRegistration = fakes.Package,
-                            NewOwner = fakes.User,
-                            ConfirmationCode = "token"
-                        });
-
-                    var policyService = new Mock<ISecurityPolicyService>();
-                    foreach (var user in usersSubscribed)
-                    {
-                        policyService.Setup(s => s.SubscribeAsync(user, "SecurePush"))
-                            .Returns(Task.FromResult(true))
-                            .Verifiable();
-                    }
-
-                    var userService = new Mock<IUserService>();
-                    userService.Setup(x => x.FindByUsername(fakes.User.Username)).Returns(fakes.User);
-
-                    var policyMessages = new Dictionary<string, string>();
-                    var messageService = new Mock<IMessageService>();
-                    messageService.Setup(s => s.SendPackageOwnerAddedNotice(
-                        It.IsAny<User>(), It.IsAny<User>(), It.IsAny<PackageRegistration>(), It.IsAny<string>(), It.IsAny<string>()))
-                        .Callback<User, User, PackageRegistration, string, string>((toUser, newOwner, pkg, pkgUrl, policyMessage) =>
-                        {
-                            policyMessages.Add(toUser.Username, policyMessage);
-                        });
-
-                    var controller = CreateController(
-                        GetConfigurationService(),
-                        httpContext: mockHttpContext,
-                        packageService: packageService,
-                        packageOwnershipManagementService: packageOwnershipManagementService,
-                        messageService: messageService,
-                        securityPolicyService: policyService,
-                        userService: userService);
-
-                    controller.SetCurrentUser(fakes.User);
-                    TestUtility.SetupHttpContextMockForUrlGeneration(mockHttpContext, controller);
-
-                    // Act
-                    await controller.ConfirmPendingOwnershipRequest(fakes.Package.Id, fakes.User.Username, "token");
-
-                    // Assert
-                    foreach (var user in usersSubscribed)
-                    {
-                        policyService.Verify(s => s.SubscribeAsync(user, "SecurePush"), Times.Once);
-                    }
-
-                    return policyMessages;
-                }
-            }
         }
 
         public class TheContactOwnersMethod
