+++ tests/NuGetGallery.Facts/Security/SecurityPolicyServiceFacts.cs
@@ -55,10 +55,9 @@ namespace NuGetGallery.Security
 
             // Assert
             Assert.NotNull(handlers);
-            Assert.Equal(3, handlers.Count);
-            Assert.Equal(typeof(RequireMinClientVersionForPushPolicy), handlers[0].GetType());
-            Assert.Equal(typeof(RequirePackageVerifyScopePolicy), handlers[1].GetType());
-            Assert.Equal(typeof(RequireMinProtocolVersionForPushPolicy), handlers[2].GetType());
+            Assert.Equal(2, handlers.Count);
+            Assert.Equal(typeof(RequirePackageVerifyScopePolicy), handlers[0].GetType());
+            Assert.Equal(typeof(RequireMinProtocolVersionForPushPolicy), handlers[1].GetType());
         }
 
         [Fact]
